feat: Enable user selection
Added methods to select characters and get/submit words, plus a number
of helper methods.

Switched to a 1D grid. This makes it slightly harder to shift letters
down later on, and to check the congruity of the letters, but it does
make it easier to deduplicate the selection. I'm still not sold on 1D
over 2D so this might change.