task: Implement basic game UI
Basic board with 10x10 grid.

This is obviously not an ideal way to generate the board tiles, but hey,
it works.

I considered rendering the tiles with js, but (aside from using a
framework) this messes with the JS and it feels unnecessary when the
board itself isn't going to be moving, just the content inside.

The CSS will need some tweaking; it's currently quite ugly.