feat: Add select/deselect/submit/reset functionality
Now able to select, deselect, submit and reset input on the game board.

It's a little hacky, but each tile element has an `onclick` event which
calls `select` with a hard-coded tile index. The `WordHandler` select
toggles the user's selection on and off. Eventually I'll add logic to
only allow the user to click contiguous tiles.

I've changed the `grid` to be a list of `Tile` objects. This should make
it a bit easier to manage and render the state of individual tiles.