refactor: Push tiles to `user_selection`
Instead of pushing tile indexes to the `user_selection` list, we can
push the tiles themselves. To my surprises, `Array.includes` is able to
detect if the `Tile` object exists in `user_selection`, so it's
functionally the same as pushing the indexes.

However it makes it much easier to manipulate the tiles themselves from
the `user_selection` array. For example when we come to do the scoring,
we won't have to access the tile through the grid - we can access all of
the tile info from the `user_selection` array.