feat: Implement scoring
Using the tile scores, I've added a scoring mechanism.

With the scores, my eventual idea is that you'll have maybe 2-5 minutes
to play the game each day, and in that time you have to achieve as high
of a score as possible. Then maybe I can store the top 5 scores with the
dates, and display them on a leaderboard to the user.