fix: Remove double-tap delay
On mobile I noticed that the tile presses had a signficant amount of
delay. This is apparently intentional; it gives the browser time to
detect double-taps. I thought it was a performance problem, so I'm glad
to learn it's a literal feature and not a bug.

This should improve the responsiveness of the tiles.

[Source](https://developers.google.com/web/updates/2013/12/300ms-tap-delay-gone-away).