improvement: Distribute letters
This ensures letters are selected based on their frequency in the
English language. Distributions from
[nd.edu](https://www3.nd.edu/~busiforc/handouts/cryptography/letterfrequencies.html).

I've included (approximately) the same amount of characters in the
`CHARS` string that that source suggests.

There are probably more intelligent ways to do this, but it's O(1)
complexity for selection and it doesn't take up that much memory.