improvement: Update button style
Instead of making the buttons look like the tiles (because that's quite
confusing), I've made the buttons look more like (wait for it) buttons.

Made the text lowercase to match the rest of the theme.

Also managed to fix the layout problem where the "clear" button was
misaligned from the grid - the outer container wasn't wide enough!