feat: Add multiplier to long words
Added a score multiplier to words longer than `MULTIPLIER_THRESHOLD` in
the constants file. If that threshold is reached, a `x2` multiplier
label appears in the score row, and the colour of the tiles change to
`TILE_MULTIPLIER_BG_COLOUR` (also in the constants file).

As part of this, I've also tidied up the styling of the score row
elements to make the styles more applicable across the application. The
HTML is a bit more structured for consistency.

I like the mechanism, but I'm uncertain of three things:

* The threshold seems a bit high. Maybe 6 would be better?
* The colour doesn't feel... like an achievement. Maybe gold?
* A 10+ word would be really hard. Maybe a 3x multiplier for that?