fix: Scale UI for mobiles
UI previously didn't scale below a certain size.

This ensures that the grid flexibly scales at small sizes by making the
tile height equal to the tile width, which is proportional to the screen
width. More [here](http://alistapart.com/article/creating-intrinsic-ratios-for-video/).

This kind of messed with the subscript due to the dynamic vertical
alignment of the tile content, but I've fixed that with some margin
tweaking.

This also creates a small margin at the bottom of the grid on safari
because 6 doesn't go into 100 perfectly, which is going to annoy me no
end, so I might change the grid to be 8x8.