feat: Store user's highest scoring word
Rough first pass at storing the user's highest submitted word. The word
has to be valid and higher scoring than their last top word.

This needs some improving but it lays the groundwork for the local
storage stuff which will be a core mechanism for the main game loop,
restricting plays over each day.

Maybe this should be loaded and unloaded into a json object when the
browser loads/closes. Much easier to manage.