feat(stats): Add modal
Added a modal for stats, which for now just displays the top word the
user has entered.

Fixed a bug where words lower than the current top word were being
stored, instead of higher words.

Again, this local storage stuff needs to be optimised as the amount
we're storing grows.