improvement(toast): Change style + fix bugs
Toast now stretches across the bottom of the screen, which looks a bit
neater. I think this could be further improved by centering the text.

Shortened the `IncongruousSelectionError` error message to better fit in
the toast.

Fixed bug with toast being hidden too often. The hide is reset if a new
toast appears.