task(buffer): Save and load
When the game starts up, the buffer is loaded from local storage if it
exists. Eventually, this will handle the buffer expiration too. This
overwrites the default buffer generated by the GitHub action.

When a word is submitted, the grid is concatenated with the buffer and
written back to local storage. This ensures the grid and buffer state is
preserved.

For now, if the buffer runs out, the tiles won't be randomised. But the
game will end long before that, so it's not a problem.