task: revert changes to generation
Apparently cron schedules can sometimes take a while to run, which is why we haven't seen this run yet.

I'm setting this to 1 in the morning to see if it happens overnight. Will revert changes if it does.

This obviously isn't ideal. This could mean users sometimes get the old grid rather than a new grid.