fix(refresh): Tie griddle ID to buffer gen
Previously the griddle IDs were generated client side.

Due to the delay in GitHub running Actions, it is possible that the user
loads the game at 00:01UTC before github has updated the buffer. This
would cause their game to update the current griddle ID to today's date,
giving them the same griddle as yesterday.

This ensures the griddle IDs are only updated when the buffer updates,
ensuring the game only refreshes when there is a new griddle.

Longer term, this absolutely needs a more reliable mechanism to run the
buffer generation.