fix(generate_buffer): Switch back to 00:00UTC
GitHub Actions that start at 00:00 (or on the hour) are often delayed
due to a high volume of Actions at this time. For griddle `2022-02-22`,
it took an hour to run the script.

Now that the griddle IDs are tied to the buffer generation, it doesn't
matter so much for the Action to be delayed. Best case scenario, it
happens as expected around 00:00UTC.